scripts/automations/requirements.txt
libsql-experimental
python-dotenv
twilio
